+++ setup.cfg
@@ -16,6 +16,7 @@ packages =
     ethereum_spec_tools
     ethereum
     ethereum/frontier
+    ethereum/frontier/utils
     ethereum/frontier/vm
     ethereum/frontier/vm/instructions
     ethereum/utils
