+++ src/ethereum/frontier/utils/__init__.py
@@ -1,0 +1,13 @@
+"""
+Frontier Utility Functions
+^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+..contents:: Table of Contents
+    :backlinks: none
+    :local:
+
+Introduction
+------------
+
+Utility functions used in this frontier version of specification.
+"""
