+++ src/ethereum/frontier/utils/address.py
@@ -1,6 +1,6 @@
 """
-Utility Functions For Addresses
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Frontier Utility Functions For Addresses
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 ..contents:: Table of Contents
     :backlinks: none
@@ -9,7 +9,7 @@ Utility Functions For Addresses
 Introduction
 ------------
 
-Address specific utility functions used in this application.
+Address specific functions used in this frontier version of specification.
 """
 from typing import Union
