+++ src/ethereum/frontier/vm/instructions/environment.py
@@ -13,8 +13,8 @@ Implementations of the EVM environment related instructions.
 """
 
 from ethereum.base_types import U256, Uint
+from ethereum.frontier.utils.address import to_address
 from ethereum.frontier.vm.memory import extend_memory, memory_write
-from ethereum.utils.address import to_address
 from ethereum.utils.numeric import ceil32
 
 from .. import Evm
