+++ tests/frontier/blockchain_st_test_helpers.py
@@ -12,6 +12,7 @@ from ethereum.frontier.eth_types import (
     State,
     Transaction,
 )
+from ethereum.frontier.rlp import rlp_hash
 from ethereum.frontier.spec import BlockChain, state_transition
 from ethereum.utils.hexadecimal import (
     hex_to_address,
@@ -23,7 +24,6 @@ from ethereum.utils.hexadecimal import (
     hex_to_u256,
     hex_to_uint,
 )
-from ethereum.utils.rlp import rlp_hash
 
 TEST_DIR = (
     "tests/fixtures/LegacyTests/Constantinople/BlockchainTests/"
